# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/crypto/analyze/sort/by/normalized/desc":{"post":{"tags":["crypto-analyzer-controller"],"summary":"Returns a list of cryptos sorted by normalized price range in descending order. Normalized range is also returned with each crypto. Cryptos for which data do not exist are not present in the list. An optional time period can be specified to restrict results. If not provided, search is not restricted and if any of its fields are not provided, searching is restricted to only the other field. Time period fields must conform to pattern \"yyyy-MM-dd HH:mm:ss\"","operationId":"sortCryptosByNormalizedRangeDesc","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimePeriodRequestDto"},"example":{"fromDateTime":"2023-01-31 20:03:44","toDateTime":"2023-02-29 23:19:01"}}}},"responses":{"500":{"description":"Internal error during crypto data loading, manipulations or calculations","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"200":{"description":"Successfully calculated list of cryptos","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CryptoMetricsListDto"}}}}}}},"/api/crypto/analyze/metrics/{crypto}":{"post":{"tags":["crypto-analyzer-controller"],"summary":"Returns a mapping of metrics to their corresponding values for the specified crypto. Mapping is empty if data do not exist for specified crypto. Metrics returned are oldest,newest,minimum,maximum and normalized range. An optional time period can be specified to restrict results. If not provided, search is not restricted and if any of its fields are not provided, searching is restricted to only the other field. Time period fields must conform to pattern \"yyyy-MM-dd HH:mm:ss\"","operationId":"getMetricsForCrypto","parameters":[{"name":"crypto","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimePeriodRequestDto"},"example":{"fromDateTime":"2023-01-31 20:03:44","toDateTime":"2023-02-29 23:19:01"}}}},"responses":{"500":{"description":"Internal error during crypto data loading, manipulations or calculations","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"Requested crypto is not supported","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"200":{"description":"Successfully calculated metrics for crypto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CryptoMetricsDto"}}}}}}},"/api/crypto/analyze/crypto/with/highest/normalized":{"post":{"tags":["crypto-analyzer-controller"],"summary":"Returns the crypto with the highest normalized range. Its normalized range value is also returned. Response contains null values if crypto data do not exist. An optional time period can be specified to restrict results. If not provided, search is not restricted and if any of its fields are not provided, searching is restricted to only the other field. Time period fields must conform to pattern \"yyyy-MM-dd HH:mm:ss\"","operationId":"getCryptoWithHighestNormalizedRange","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimePeriodRequestDto"},"example":{"fromDateTime":"2023-01-31 20:03:44","toDateTime":"2023-02-29 23:19:01"}}}},"responses":{"500":{"description":"Internal error during crypto data loading, manipulations or calculations","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"200":{"description":"Successfully found crypto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CryptoMetricDto"}}}}}}}},"components":{"schemas":{"TimePeriodRequestDto":{"type":"object","properties":{"fromDateTime":{"type":"string","format":"date-time"},"toDateTime":{"type":"string","format":"date-time"}}},"ErrorResponseDto":{"type":"object","properties":{"errorMessage":{"type":"string"}}},"CryptoMetricDto":{"type":"object","properties":{"crypto":{"type":"string"},"value":{"type":"number"}}},"CryptoMetricsListDto":{"type":"object","properties":{"cryptos":{"type":"array","items":{"$ref":"#/components/schemas/CryptoMetricDto"}}}},"CryptoMetricsDto":{"type":"object","properties":{"metrics":{"type":"object","additionalProperties":{"type":"number"}}}}}}}